Runtime: 0 ms, faster than 100.00% of Java online submissions for Maximum Depth of Binary Tree. Memory Usage: 42.8 MB, less than 72.20% of Java online submissions for Maximum Depth of Binary Tree.
